partially refactored product import
